requirements.txt
numpy>=1.26
scipy>=1.12
pandas>=2.2
networkx>=3.2
trimesh>=4.4
PyYAML>=6.0
python-dotenv>=1.0
tqdm>=4.66
caveclient>=5.20
fafbseg>=3.2
navis>=1.7
meshparty>=1.18
cloud-volume>=11.0
